Make heater use pid
Former-commit-id: f57c56d93c82de6f9f005623d0b9cd3c08a6a4ac

+++ src/core/src/grid/grid.rs
@@ -18,7 +18,7 @@ pub struct Electrode {
 #[derive(Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Clone, Copy)]
 #[serde(tag = "type")]
 pub enum Peripheral {
-    Heater { pwm_channel: u32, spi_channel: u8 },
+    Heater { pwm_channel: u8, spi_channel: u8 },
     Input,
 }
